Add files via upload
This is a famous cup and ball game.
There are three cups and a ball. The ball is under one cup.
First, the ball is shown under the cup, where the position is visible.
Then, the cups are shuffled. 
Now you have to guess the ball is under which cup numbering  0,1,2.